Round 2: dedupe value serializer, split deluge event model, clear warnings
- values.rs: value_to_string and value_to_json_string were byte-identical;
  collapse to a single value_to_json_string and repoint callers.
- deluge/event.rs: extract DelugeEvent + parse_event out of the transport
  file, separating the event domain model from the wire/connection code.
- Document why several DelugeEvent payload fields are decoded-but-unread
  (surfaced via Debug in the event log) with a single explained
  #[allow(dead_code)], clearing the 7 long-standing dead-field warnings.

Build is now warning-free; all 40 tests pass.
